Change package name
Previously, the package name was still duke, didn't make any sense.

Now, package name reflects the actual name of the bot.
