+++ build.gradle
@@ -41,7 +41,7 @@ test {
 }
 
 application {
-    mainClassName = "duke.Launcher"
+    mainClassName = "chibot.Launcher"
 }
 
 shadowJar {
