--- src/main/java/duke/commands/HelpCommand.java
@@ -1,18 +1,0 @@
-package duke.commands;
-
-import duke.storage.Storage;
-import duke.tasklist.TaskList;
-
-public class HelpCommand extends Command {
-
-    private final String[] tokens;
-
-    public HelpCommand(String[] tokens) {
-        this.tokens = tokens;
-    }
-
-    @Override
-    public String execute(TaskList tl, Storage sge) {
-        return "";
-    }
-}
