--- src/main/java/duke/storage/StorageStub.java
@@ -1,16 +1,0 @@
-package duke.storage;
-
-/**
- * A STUB for testing classes with Storage class dependencies.
- */
-public class StorageStub extends Storage {
-
-    /**
-     * Constructor of the class.
-     *
-     * @param s Filler String to be passed into parent class.
-     */
-    public StorageStub(String s) {
-        super(s);
-    }
-}
